examples/gpt4-code-interpreter-py/pyproject.toml
[tool.poetry]
name = "gpt4-code-interpreter"
version = "0.1.0"
description = ""
authors = ["Vasek Mlejnsky <vasek.mlejnsky@gmail.com>"]
readme = "README.md"
packages = [{include = "gpt4_code_interpreter"}]

[tool.poetry.dependencies]
python = "^3.10"
e2b = "^0.4.1"
openai = "^0.28.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
